Lab5/requirements.txt
gradio>=4.19.0
azure-search-documents>=11.4.0
azure-core>=1.30.0
openai>=1.12.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
azure-storage-blob>=12.0.0
python-multipart>=0.0.9
